:arrows_clockwise: Switch order days are checked.
Change order of which days are inspected to find a matching opening times session (was check yesterday, for midnight spanning opening times, then today). Now checks today before checking yesterday. Did this on the basis that opening times which span midnight are rare so most checks would end up checking yesterday and today if checking yesterday first. This shaved 50 ms of the timings.

․ timings isOpen for 1000 orgs disregarding alterations: 463ms
․ timings isOpen for 1000 orgs including alterations: 678ms