:performing_arts: Return open/closed status object
Rather than a simple isOpen method return an object containing 'isOpen'
flag as well as 'until' value. Method renamed getStatus to reflect this.

This is a performance optimisation based on the observation that we
often want both bits of information and can calculate them at the same
time. Will probably introduce a paramater to control whether 'until' is
calculated.

Rationalised and cleaned up the tests since there was quite a bit of duplication